external/atria/pyproject.toml
[build-system]
requires = [
    "setuptools>=60",
    "setuptools-scm>=8.0",
    "wheel",
    "pip==24.3.1",
]
build-backend = "setuptools.build_meta"

[project]
name = "atria"
version = "0.0.0"                                                                                                    
description = "atria is a PyTorch-based toolkit designed for training/testing machine/deep learning models at scale."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE.txt" }
keywords = ["torch", "machine-learning", "deep-learning"]
authors = [{ name = "Saifullah", email = "saifullah.saifullah@dfki.de" }]
maintainers = [{ name = "Saifullah", email = "saifullah.saifullah@dfki.de" }]

# Classifiers help users find your project by categorizing it.
# For a list of valid classifiers, see https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
    "autopep8==2.1.0",
    "beautifulsoup4==4.12.3",
    "bs4==0.0.2",
    "click==8.0.1",
    "coloredlogs==15.0.1",
    "dacite==1.6.0",
    "dataclasses==0.6",
    "datasets==2.19.1",
    "diffusers==0.31.0",
    "einops==0.8.0",
    "hydra-core==1.3.2",
    "lxml==5.2.1",
    "omegaconf==2.3.0",
    "opencv-python==4.8.0.74",
    "optuna==3.6.1",
    "pandas",
    "Pillow==10.0.0",
    "protobuf==5.26.1",
    "pycocotools",
    "pycodestyle==2.11.1",
    "pynvml==11.5.0",
    "pytorch-fid==0.3.0",
    "pytorch-ignite==0.5.0.post2",
    "scikit-image==0.24.0",
    "scikit-learn==1.5.2",
    "scipy==1.14.1",
    "seqeval==1.2.2",
    "tensorboardX==2.6.2.2",
    "timm==0.9.16",
    "torchvision==0.18.0",
    "tqdm==4.66.4",
    "transformers==4.40.2",
    "anls==0.0.2",
    "textdistance==4.6.3",
    "h5py==3.12.1",
    "hydra_zen==0.13.0",
    "webdataset==0.2.100",
    "pydantic==2.9.2",
    "pytesseract==0.3.13",
    "more_itertools==10.5.0",
    "ray==2.38.0",
    "ocrodeg @ git+https://github.com/NVlabs/ocrodeg.git",
    "datadings @ git+https://github.com/saifullah3396/datadings.git",
]

[project.optional-dependencies]
test = ["coverage"]

[project.urls]
"Homepage" = "https://github.com/saifullah3396/atria/"
"Bug Reports" = "https://github.com/saifullah3396/atria/"
"Source" = "https://github.com/saifullah3396/atria/"

[tool.setuptools.packages.find]
where = ["src"]
include = ["atria*"]

[tool.setuptools.package-data]
atria = ["*.txt", "*.yaml"]

[tool.pytest.ini_options]
pythonpath = [".", "src", "tests"]

[tool.distutils.egg_info]
egg_base = "./"

[project.scripts]
"atria.prepare_data" = "atria.core.task_runners.atria_data_processor:app"
"atria.evaluate" = "atria.core.task_runners.atria_model_evaluator:app"
"atria.train" = "atria.core.task_runners.atria_trainer:app"
